refactoring kanban
