+++ ScrumEnv/src/ProjectBundle/Controller/DefaultController.php
@@ -150,77 +150,5 @@ class DefaultController extends Controller
     }
 
 
-     public function visualisationAction($owner, $project){
-        $em = $this->container->get('doctrine')->getEntityManager();
-
-        $US= $em->getRepository('ProjectBundle:Task')->findBy(
-            array('owner' => $owner,
-                'project' => $project));
-
-        return $this->container->get('templating')->renderResponse('ProjectBundle:Kanban:Kanban_visualisation.html.twig', 
-        array(
-        'kanban_us' => $US,
-        'owner' => $owner,
-        'project' => $project
-        ));
-    }
-
-    public function Update_task_achievementoGAction($owner, $project, $id){
-
-         $message = '';
-         $up = 'ok';
-         $em = $this->container->get('doctrine')->getEntityManager();
-         $Task = $em->find('ProjectBundle:Task', $id);
-         $US= $em->getRepository('ProjectBundle:Task')->findAll();
-
-
-         if (!$Task){
-            $message = "Aucune tache trouve";
-         }
-
-         else{
-            $Task -> setAchievementTask("onGoing");
-            $em->persist($Task);
-            $em->flush();
-         }
-          return $this->container->get('templating')->renderResponse('ProjectBundle:Kanban:Kanban_visualisation.html.twig',array('kanban_us' => $US, 'message' => $message, 'up' => $up,'owner' => $owner, 'project' => $project));
-    }
-
-    public function Update_task_achievementDAction($owner, $project, $id){
-
-         $message = '';
-         $up = 'ok';
-         $em = $this->container->get('doctrine')->getEntityManager();
-         $Task = $em->find('ProjectBundle:Task', $id);
-          $US= $em->getRepository('ProjectBundle:Task')->findAll();
-         if (!$Task){
-            $message = "Aucune tache trouve";
-         }
-
-         else{
-            $Task -> setAchievementTask("Done");
-            $em->persist($Task);
-            $em->flush();
-         }
-          return $this->container->get('templating')->renderResponse('ProjectBundle:Kanban:Kanban_visualisation.html.twig',array('kanban_us' => $US, 'message' => $message, 'up' => $up,'owner' => $owner, 'project' => $project));
-    }
-
-    public function Update_task_achievementToDoAction($owner, $project, $id){
-
-         $message = '';
-         $up = 'ok';
-         $em = $this->container->get('doctrine')->getEntityManager();
-         $Task = $em->find('ProjectBundle:Task', $id);
-          $US= $em->getRepository('ProjectBundle:Task')->findAll();
-         if (!$Task){
-            $message = "Aucune tache trouve";
-         }
-
-         else{
-            $Task -> setAchievementTask("ToDo");
-            $em->persist($Task);
-            $em->flush();
-         }
-          return $this->container->get('templating')->renderResponse('ProjectBundle:Kanban:Kanban_visualisation.html.twig',array('kanban_us' => $US, 'message' => $message, 'up' => $up,'owner' => $owner, 'project' => $project));
-    }
+     
 }
